MDL-16075 File storage conversion of Glossary

+++ mod/glossary/deleteentry.php
@@ -3,17 +3,16 @@
     require_once("../../config.php");
     require_once("lib.php");
 
-    $id      = required_param('id', PARAM_INT);          // course module ID
-    $confirm = optional_param('confirm', 0, PARAM_INT);  // commit the operation?
-    $entry   = optional_param('entry', 0, PARAM_INT);    // entry id
+    $id       = required_param('id', PARAM_INT);          // course module ID
+    $confirm  = optional_param('confirm', 0, PARAM_INT);  // commit the operation?
+    $entry    = optional_param('entry', 0, PARAM_INT);    // entry id
+    $prevmode = required_param('prevmode', PARAM_ALPHA);
+    $hook     = optional_param('hook', '', PARAM_CLEAN);
 
-    $prevmode = required_param('prevmode');
-    $hook = optional_param('hook', '', PARAM_CLEAN);
-
-    $strglossary = get_string("modulename", "glossary");
+    $strglossary   = get_string("modulename", "glossary");
     $strglossaries = get_string("modulenameplural", "glossary");
-    $stredit = get_string("edit");
-    $entrydeleted = get_string("entrydeleted","glossary");
+    $stredit       = get_string("edit");
+    $entrydeleted  = get_string("entrydeleted","glossary");
 
 
     if (! $cm = get_coursemodule_from_id('glossary', $id)) {
@@ -30,8 +29,8 @@
 
     require_login($course->id, false, $cm);
     $context = get_context_instance(CONTEXT_MODULE, $cm->id);
-    $manageentries = has_capability('mod/glossary:manageentries', $context); 
-    
+    $manageentries = has_capability('mod/glossary:manageentries', $context);
+
     if (! $glossary = $DB->get_record("glossary", array("id"=>$cm->instance))) {
         print_error('invalidid', 'glossary');
     }
@@ -40,9 +39,6 @@
     $strareyousuredelete = get_string("areyousuredelete","glossary");
 
     $navigation = build_navigation('', $cm);
-    print_header_simple(format_string($glossary->name), "", $navigation,
-                  "", "", true, update_module_button($cm->id, $course->id, $strglossary),
-                  navmenu($course, $cm));
 
     if (($entry->userid != $USER->id) and !$manageentries) { // guest id is never matched, no need for special check here
         print_error('nopermissiontodelentry');
@@ -54,22 +50,40 @@
 
 /// If data submitted, then process and store.
 
-    if ($confirm) { // the operation was confirmed.
+    if ($confirm and confirm_sesskey()) { // the operation was confirmed.
         // if it is an imported entry, just delete the relation
 
-        if ( $entry->sourceglossaryid ) {
-            $dbentry = new stdClass;
-            $dbentry->id = $entry->id;
-            $dbentry->glossaryid = $entry->sourceglossaryid;
-            $dbentry->sourceglossaryid = 0;
-            if (! $DB->update_record('glossary_entries', $dbentry)) {
+        if ($entry->sourceglossaryid) {
+            if (!$newcm = get_coursemodule_from_instance('glossary', $entry->sourceglossaryid)) {
+                print_error('invalidcoursemodule');
+            }
+            $newcontext = get_context_instance(CONTEXT_MODULE, $newcm->id);
+
+            $entry->glossaryid       = $entry->sourceglossaryid;
+            $entry->sourceglossaryid = 0;
+            if (!$DB->update_record('glossary_entries', $entry)) {
                 print_error('cantupdateglossary', 'glossary');
             }
 
-        } else {
-            if ( $entry->attachment ) {
-                glossary_delete_old_attachments($entry);
+            // move attachments too
+            $fs = get_file_storage();
+
+            if ($oldfiles = $fs->get_area_files($context->id, 'glossary_attachment', $entry->id)) {
+                foreach ($oldfiles as $oldfile) {
+                    $file_record = new object();
+                    $file_record->contextid = $newcontext->id;
+                    $fs->create_file_from_storedfile($file_record, $oldfile);
+                }
+                $fs->delete_area_files($context->id, 'glossary_attachment', $entry->id);
+                $entry->attachment = '1';
+            } else {
+                $entry->attachment = '0';
             }
+            $DB->update_record('glossary_entries', $entry);
+
+        } else {
+            $fs = get_file_storage();
+            $fs->delete_area_files($context->id, 'glossary_attachment', $entry->id);
             $DB->delete_records("glossary_comments", array("entryid"=>$entry->id));
             $DB->delete_records("glossary_alias", array("entryid"=>$entry->id));
             $DB->delete_records("glossary_ratings", array("entryid"=>$entry->id));
@@ -77,16 +91,20 @@
         }
 
         add_to_log($course->id, "glossary", "delete entry", "view.php?id=$cm->id&amp;mode=$prevmode&amp;hook=$hook", $entry->id,$cm->id);
-        redirect("view.php?id=$cm->id&amp;mode=$prevmode&amp;hook=$hook", $entrydeleted);
+        redirect("view.php?id=$cm->id&amp;mode=$prevmode&amp;hook=$hook");
 
     } else {        // the operation has not been confirmed yet so ask the user to do so
+        print_header_simple(format_string($glossary->name), "", $navigation,
+                      "", "", true, update_module_button($cm->id, $course->id, $strglossary),
+                      navmenu($course, $cm));
+        $areyousure = "<b>".format_string($entry->concept)."</b><p>$strareyousuredelete</p>";
+        $linkyes    = 'deleteentry.php';
+        $linkno     = 'view.php';
+        $optionsyes = array('id'=>$cm->id, 'entry'=>$entry->id, 'confirm'=>1, 'sesskey'=>sesskey(), 'prevmode'=>$prevmode, 'hook'=>$hook);
+        $optionsno  = array('id'=>$cm->id, 'mode'=>$prevmode, 'hook'=>$hook);
 
-        notice_yesno("<b>".format_string($entry->concept)."</b><p>$strareyousuredelete</p>",
-                      "deleteentry.php?id=$cm->id&amp;mode=delete&amp;confirm=1&amp;entry=".s($entry->id)."&amp;prevmode=$prevmode&amp;hook=$hook",
-                      "view.php?id=$cm->id&amp;mode=$prevmode&amp;hook=$hook");
+        notice_yesno($areyousure, $linkyes, $linkno, $optionsyes, $optionsno, 'post', 'get');
 
+        print_footer($course);
     }
-
-    print_footer($course);
-
 ?>
